API - Filtrado de fotos en función del usuario registrado (get_queryset con reglas personalizadas, las mismas que las de la web). Checkpermissions para GET, POST, PUT y DELETE users.

+++ Frikr/users/api.py
@@ -11,6 +11,7 @@ from rest_framework import status
 from rest_framework.pagination import PageNumberPagination
 
 from users.serializers import UserSerializer
+from users.permissions import UserPermission
 
 # FORMA TRADICIONAL con vistas de Django en lugar de APIView
 # class UserListAPI(View):
@@ -30,7 +31,10 @@ from users.serializers import UserSerializer
 # Se encarga él de devolver lo que corresponda. Además de proporcionar una vista de presentación.
 class UserListAPI(APIView):
 
+    permission_classes = (UserPermission,)
+
     def get(self, request):
+        self.check_permissions(request)
         users = User.objects.all()
         # Como nos traemos todos los objetos del tirón, la paginación no es automática tras añadir
         # la variable REST_FRAMEWORK en los settings. Hay que hacer lo siguiente para la paginación
@@ -50,6 +54,7 @@ class UserListAPI(APIView):
         return paginator.get_paginated_response(serialized_user)
 
     def post(self, request):
+        self.check_permissions(request)
         # OJO!! Con REST, te guardará los datos en 'data', en lugar de POST, etc.
         serializer = UserSerializer(data=request.data)
         if serializer.is_valid():
@@ -61,16 +66,26 @@ class UserListAPI(APIView):
 
 class UserDetailAPI(APIView):
 
+    permission_classes = (UserPermission,)
+
     def get(self, request, pk):
+        self.check_permissions(request)
+
         # Django nos proporciona este método común que nos devuelve el objeto de clase User con filtro pk=pk.
         # Si no lo encuentra, lanza una excepción 404-Not Found que captura el propio Django.
         user = get_object_or_404(User, pk=pk)
+
+        self.check_object_permissions(request, user)
 
         serializer = UserSerializer(user)
         return Response(serializer.data)
 
     def put(self, request, pk):
+        self.check_permissions(request)
+
         user = get_object_or_404(User, pk=pk)
+
+        self.check_object_permissions(request, user)
 
         serializer = UserSerializer(instance=user, data=request.data)
         if serializer.is_valid():
@@ -80,6 +95,11 @@ class UserDetailAPI(APIView):
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
+        self.check_permissions(request)
+
         user = get_object_or_404(User, pk=pk)
+
+        self.check_object_permissions(request, user)
+
         user.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
